migrate ciphering text to self modifying component, fix active letter classes
