+++ react/src/docs/react/ciphering-text/Page.module.css
@@ -1,12 +1,12 @@
-pre {
+.ciphering-text {
 	font-size: 24px;
 	line-height: 54px;
 }
 
-pre > span {
+.ciphering-text__span {
 	transition: all 50ms ease-in-out;
 }
 
-pre > span.active {
+.ciphering-text__span.active {
 	color: #e3caca;
 }
